backend/note-service/requirements.txt
# FastAPI ve Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# JWT Authentication ← YENİ
PyJWT==2.8.0

# Password Hashing ← YENİ (login için gerekli)
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0
httpx==0.27.0

# Validation
pydantic==2.9.0
pydantic-settings==2.5.0
pydantic[email]==2.9.0

# Optional
# alembic==1.13.0  # Database migrations